AI_Career_Skills_Advisor/AI_Career_Skills_Advisor/requirements.txt

streamlit
pandas
spacy
PyPDF2
python-docx
plotly
